dev-1.2.2.77
Added the Callback Wildcard.
Activates nerfed versions of all the wildcards in a random order throughout the session. The session starts with a new wildcard every ~5 minutes, and max one wildcard at once. As the session goes on, the time between activating wildcards decreases, and the maximum amount of allowed wildcards increases too. Near the end of the session (5 minutes before the 3/4 mark), all wildcards are activated for the next five minutes. The remaining quarter of the session does not have any wildcards active.

Callback is obviously supposed to be the last wildcard on the series finale, so it will have to be MANUALLY activated (it will not be included in the wildcard selection when the session starts [This still is not implemented yet btw]).