dev-1.2.2.83
Added the '/wildcard choose' command, which opens a gui that lets you choose the wildcard that will be used that session.
The same gui also opens when you starta session without selecting a wildcard beforehand. If you still don't select it, a random wildcard will be selected for you (excluding Callback)