1.3.4.1
The game no longer crashes when someone is using the astral projection power and a player joins or leaves the server.
Synced all timers so that they tick down at once (this is only visual, no actual behavior was changed).
You can no longer mimic a player with the mimicry power (unless that player has already mimicked someone else).
If a person joins late to a session with the Superpowers wildcard active, they will automatically be assigned a superpower.