dev-1.3.4.3
Added a command to get the lives of everyone (/lives get *).
If the players_drop_items_on_final_death config is turned on, the inventory of the players that die will now be cleared to prevent duplicating the items.
Zombie deaths no longer spam the chat with the 'Unjustified kill' messages.
When zombies are killed, their lives can now go into the negatives.
Crash fix (maybe? idk)