dev-1.3.4.6
Added the 'tick_offline_players' config for Limited Life: Controls whether even players that are offline lose time when the session is on. Default value is false.

Any items with the NoAnvil component now can't be used in the anvil.
Any items with the NoCrafting component now can't be used in the crafting menu.
Any items with the NoEnchants component now can't be used in the enchantment table.
Any items with the NoModifications component act as if they had all three of the listed above - NoAnvil, NoCrafting, NoEnchants.
The trivia / Secret Life mace now has the NoModifications component, instead of the previous one it had (NoMending).