1.3.5
Fixed issues with the Limited Life client-side lives timer desyncyng from the server-side one.
This also unfortunately means that the lives timer no longer ticks down at the same timer as the session timer, but what can ya do.
Temporarily disabled the Trivia Bot display in the quiz gui, as it's not working properly in 1.21.6.
Updated Stonecutter to 0.7-beta.4