dev-1.3.5.2
Added a way to set a specific players' task to what you want, using the `/task set player <task_difficulty> <task>` command.
If the player already has a task book, it will be instantly replaced with the new one. If not, this command will force their task to be what you set it to on the next randomization.
So for example: `/task set Mat0u5 easy Open the end portal.` would give the player Mat0u5 the easy task book which says "Open the end portal".