dev-1.3.5.6
The session now properly stops when you change a series, or when the server stops.
Added the 'wildcard_superpowers_superspeed_step' Wild Life config, which controls whether you can step up blocks when using the superspeed power (like riding a horse). Default value is `false`.
When zombies fully revive (if the config allows it), their health is now reset to normal, and they are no longer treated still in the zombies list.
Fixed a bug with necromancy + mimic.