dev-1.3.5.8
**Renamed all the usages of 'series' to 'season'.**

The only place 'series' still remains is if it refers to the mod, Life Series, and in the three commands that choose the season you want to play, so 'chooseSeries', 'setSeries' and 'getSeries'. The reason i kept the commands the same is because i can already see the amount of people saying the mod is broken if i'd change it :)

More folder management.
{skip_build}