dev-1.3.5.24
Removed the 'auto_set_worldborder' config, which was replaced with the 'worldborder_size' integer config. Default value is 500.
Added the 'session_timer' client-side config, which serves as a replacement for the '/session timer showDisplay' command, although that option is still also available. Default value is true for all seasons except Limited Life
Added the 'colorblind_support' client-side config, which shows the team color name next to players' usernames in the tab list and above their heads in game.