dev-1.3.6.14
Added the 'announce_soulmates' config to Double Life, which shows the soulmate name when they are rolled. Default value is false.
Added the 'boogeyman_announce_outcome' global config, which broadcasts a message in chat when the boogeyman fails or is cured. Default value is true.
When the session ends and the boogeyman is offline, it now actually removes their lives instead of telling the adming to do it when the player logs on again.
Killing a boogeyman in Double Life no longer cures them (they used to be cured because technically they caused the death of their soulmate)