dev-1.3.6.16
In Double Life, all soulmates are now given a tag, `soulmate_<index>` where the index is determined by sorting all players by UUID and assigning consecutive numbers to each pair (starting from 1).
Wild Life zombies are periodically checked to be alive, and if so, they are internally removed from the zombies list and given back the default number of hearts.
Added the 'tab_list_show_exact_lives' config, which controls whether to show '4+' or the exact number of lives in the tab list when above four. Default value is false.