dev-1.3.6.21
Added the Watchers.
Watchers are players that are online, but are not affected by most season mechanics. They can only observe.
This is very useful for spectators and for admins.

Added the '/watcher' command. Subcommands: '/watcher info/list/add/remove'.

This is an experimental feature, report any bugs you find! (There will probably be many bugs.)