dev-1.3.6.34
When a player dies or logs off, their superpower now gets deactivated.

Reworked the backend logic for if a player is allowed to attack another player.
In seasons where you gain a reward for killing, you no longer do when it's an unjustified kill.
Added the 'allow_self_defense' global config, which controls whether self-defense kills should count as unjustified. Default value is true. You do not gain rewards for killing for self-defence.